+++ src/components/diff-viewer/utils/diff-helpers.ts
@@ -1,4 +1,4 @@
-import type { GitDiff, GitDiffLine, GitHunk } from "@/utils/git";
+import type { GitDiff, GitDiffLine, GitHunk } from "@/version-control/git/controllers/git";
 import type { ParsedHunk } from "./types";
 
 export const createGitHunk = (
