+++ src/components/layout/main-layout.tsx
@@ -9,7 +9,7 @@ import { usePersistentSettingsStore } from "@/settings/stores/persistent-setting
 import { useSettingsStore } from "@/settings/stores/settings-store";
 import { useBufferStore } from "@/stores/buffer-store";
 import { useUIState } from "@/stores/ui-state-store";
-import { type GitHunk, stageHunk, unstageHunk } from "../../utils/git";
+import { type GitHunk, stageHunk, unstageHunk } from "../../version-control/git/controllers/git";
 import AIChat from "../ai-chat/ai-chat";
 import BottomPane from "../bottom-pane";
 import CommandBar from "../command/components/command-bar";
