+++ src/file-system/controllers/store.ts
@@ -12,10 +12,10 @@ import { useFileTreeStore } from "@/file-explorer/controllers/file-tree-store";
 import { useBufferStore } from "@/stores/buffer-store";
 import { useProjectStore } from "@/stores/project-store";
 import { useSidebarStore } from "@/stores/sidebar-store";
-import { getGitStatus } from "@/utils/git";
-import { gitDiffCache } from "@/utils/git-diff-cache";
-import { isDiffFile, parseRawDiffContent } from "@/utils/git-diff-parser";
 import { createSelectors } from "@/utils/zustand-selectors";
+import { getGitStatus } from "@/version-control/git/controllers/git";
+import { gitDiffCache } from "@/version-control/git/controllers/git-diff-cache";
+import { isDiffFile, parseRawDiffContent } from "@/version-control/git/controllers/git-diff-parser";
 import { useGitStore } from "@/version-control/git/controllers/git-store";
 import type { FileEntry } from "../models/app";
 import type { FsActions, FsState } from "../models/interface";
