+++ src/version-control/git/controllers/git-diff-cache.ts
