+++ src/version-control/git/views/git-commit-panel.tsx
@@ -2,7 +2,7 @@ import { AlertCircle, GitCommit as GitCommitIcon, Send } from "lucide-react";
 import type React from "react";
 import { useState } from "react";
 import { cn } from "@/utils/cn";
-import { commitChanges } from "@/utils/git";
+import { commitChanges } from "@/version-control/git/controllers/git";
 
 interface GitCommitPanelProps {
   stagedFilesCount: number;
